Added pattern Result for project

+++ FlowBoard.Application/Features/Lists/Commands/DeleteList/DeleteListCommandHandler.cs
@@ -1,9 +1,10 @@
 using FlowBoard.Application.Abstractions;
+using FluentResults;
 using MediatR;
 
 namespace FlowBoard.Application.Features.Lists.Commands.DeleteList;
 
-public class DeleteListCommandHandler : IRequestHandler<DeleteListCommand, bool>
+public class DeleteListCommandHandler : IRequestHandler<DeleteListCommand, Result<bool>>
 {
     private readonly IUnitOfWorkFactory _uowFactory;
 
@@ -12,24 +13,33 @@ public class DeleteListCommandHandler : IRequestHandler<DeleteListCommand, bool>
         _uowFactory = uowFactory;
     }
 
-    public async Task<bool> Handle(DeleteListCommand command, CancellationToken cancellationToken)
+    public async Task<Result<bool>> Handle(DeleteListCommand command, CancellationToken cancellationToken)
     {
         using var uow = _uowFactory.Create();
         try
         {
-            var board = await uow.BoardRepository.GetByIdAsync(command.BoardId) ?? throw new KeyNotFoundException ("Board not found");
+            var board = await uow.BoardRepository.GetByIdAsync(command.BoardId);
+            if (board is null)
+            {
+                return Result.Fail("Board not found");
+            }
+            
             var isMember = await uow.BoardRepository.IsMemberAsync(command.BoardId, command.CurrentUserId);
-
             if (!isMember && board.CreatedBy != command.CurrentUserId)
             {
-                throw new UnauthorizedAccessException("You don't have access to this board");
+                return Result.Fail("You don't have access to this board");
             }
 
-            var list = await uow.ListRepository.GetByIdAsync(command.ListId) ?? throw new KeyNotFoundException("List not found");
+            var list = await uow.ListRepository.GetByIdAsync(command.ListId);
+            if (list is null)
+            {
+                return Result.Fail("List is not found");
+            }
 
             var result = await uow.ListRepository.DeleteAsync(list);
             uow.Commit();
-            return result;
+
+            return Result.Ok(result);
         }
         catch
         {
